+++ src/index.js
@@ -33,3 +33,4 @@ export {default as SQFormResetInitialValuesButton} from './components/SQForm/SQF
 export {default as SQFormTextarea} from './components/SQForm/SQFormTextarea';
 export {default as SQFormTextField} from './components/SQForm/SQFormTextField';
 export {default as SQFormMultiSelect} from './components/SQForm/SQFormMultiSelect';
+export {default as SQFormMaskedTextField} from './components/SQForm/SQFormMaskedTextField';